refactor(logger): Style changes
