+++ README.md
@@ -148,21 +148,22 @@ const env = {
   'AWS_ACCESS_KEY_SECRET': undefined,
   'AWS_REGION': undefined
 
-  // Kinesis configuration
+  // Kinesis configuration (single stream)
   'LOG_TO_KINESIS': undefined, // Kinesis stream name
   'LOG_TO_KINESIS_LEVEL': 'info', // log level for Kinesis
-  'LOG_TO_KINESIS_LOG_TYPE': undefined, //Bunyan stream type
+  'LOG_TO_KINESIS_LOG_TYPE': undefined, // bunyan stream type
   'KINESIS_OBJECT_MODE': true,
   'KINESIS_TIMEOUT': 5,
   'KINESIS_LENGTH': 50,
 
-  //optional if you want a pool of kinesis for failover
+  // Kinesis configuration (pool of streams for failover)
   'KINESIS_POOL': [
     {
-      //IF any of this config options are undefined will take root level
+      // if any of this config options are undefined will take root level,
+      // if exists
       'LOG_TO_KINESIS': undefined, // Kinesis stream name
       'LOG_TO_KINESIS_LEVEL': 'info', // log level for Kinesis
-      'LOG_TO_KINESIS_LOG_TYPE': undefined, //Bunyan stream type
+      'LOG_TO_KINESIS_LOG_TYPE': undefined, // bunyan stream type
       'AWS_ACCESS_KEY_ID': undefined,
       'AWS_ACCESS_KEY_SECRET': undefined,
       'AWS_REGION': undefined
