+++ lib/logger.js
@@ -24,29 +24,32 @@ module.exports = function getLogger(pkg, env, serializers) {
     var keepAliveAgent = require('./keep_alive_agent')(env);
     var KinesisWritable = require('aws-kinesis-writable');
     var KinesisStreamPool = require('aws-kinesis-writable').pool;
+    var utils = require('./utils');
 
     var stream;
-    // build kinesis stream array
     if (env.KINESIS_POOL) {
+      // pool for failover
       var kinesisStreams = env.KINESIS_POOL.map(function(configUnit) {
-        //prioritize pool config over global but if not in pool take global, ie AWS_REGION
-        return new KinesisWritable(_buildKinesisOptions(Object.assign({}, env, configUnit)), keepAliveAgent);
+        // prioritize pool config over global but if not in pool, take global (e.g. AWS_REGION)
+        return new KinesisWritable(utils.buildKinesisOptions(Object.assign({}, env, configUnit)), keepAliveAgent);
       });
       stream = new KinesisStreamPool({
         streams: kinesisStreams
       });
-    //backward compatibility writable    
     } else {
-      stream = new KinesisWritable(_buildKinesisOptions(env), keepAliveAgent);
+      // single stream, no failover
+      stream = new KinesisWritable(utils.buildKinesisOptions(env), keepAliveAgent);
     }
 
-    stream.on('error', function(err) {
+    var streamErrorHandler = function(err) {
       console.error(JSON.stringify({
         message: err.message,
         records: err.records,
         stack: err.stack
       }));
-    });
+    };
+    stream.on('error', streamErrorHandler);
+    stream.on('poolFailure', streamErrorHandler);
 
     bunyan_streams.push({
       name: 'kinesis',
@@ -54,7 +57,7 @@ module.exports = function getLogger(pkg, env, serializers) {
       level: env.LOG_TO_KINESIS_LEVEL,
       type: env.LOG_TO_KINESIS_LOG_TYPE
     });
-    
+
   }
 
   var logger = bunyan.createLogger({
@@ -71,25 +74,3 @@ module.exports = function getLogger(pkg, env, serializers) {
 
   return logger;
 };
-
-function _buildKinesisOptions(configMap, keepAliveAgent) {
-  return {
-    accessKeyId: configMap.AWS_ACCESS_KEY_ID,
-    secretAccessKey: configMap.AWS_ACCESS_KEY_SECRET,
-    streamName: configMap.LOG_TO_KINESIS,
-    region: configMap.AWS_KINESIS_REGION || configMap.AWS_REGION,
-    getCredentialsFromIAMRole: configMap.AWS_GET_CREDENTIALS_FROM_IAM_ROLE,
-    httpOptions: {
-      agent: keepAliveAgent('aws-kinesis')
-    },
-    objectMode: typeof configMap.KINESIS_OBJECT_MODE !== 'undefined' ? configMap.KINESIS_OBJECT_MODE : true,
-    buffer: {
-      timeout: configMap.KINESIS_TIMEOUT || 5,
-      length: configMap.KINESIS_LENGTH || 50,
-      isPrioritaryMsg: function (entry) {
-        return entry.level >= 40;
-      }
-    }
-  };
-}
-
